fix(ml): restore CalibrationLayer import in training script
