+++ docs/tasktracker.md
@@ -24,3 +24,11 @@
   - [ ] Добавить калибровку и отчёт по ECE
   - [ ] Записать прогнозы в БД
 - **Зависимости**: λ_final
+
+## Задача: Рефакторинг ML-слоёв
+- **Статус**: Завершена
+- **Описание**: Вынести базовые λ, модификаторы, калибровку и симулятор в ml/*.
+- **Шаги выполнения**:
+  - [x] Создать base_poisson_glm, modifiers_model, calibration, montecarlo_simulator
+  - [x] Обновить импорты в services, scripts и telegram
+- **Зависимости**: Project.md (разделы 3-4)
